+++ examples/main.py
@@ -4,17 +4,20 @@
 from chinodeco import *
 from chinodeco.decodsl import when
 
-@when(0, elsedeco = decochain(addsuffix(("[.]", 0))))(
-    addsuffix(("[_]", 0))
+@when(0)(
+    addsuffix(("[_]", "msg"))
+).elsedeco(
+    decochain(addsuffix(("[.]", 0)), addprefix(("[above]", "msg")))
 )
 @when(1)(
-    addprefix(("[system]", 0))
+    addprefix(("[system]", "msg"))
 )
 def hello(msg):
     print(msg)
 
 def main():
     hello("你好")
+    print(hello.__name__)
 
 if __name__ == "__main__":
     main()